Inference/requirements.txt
torch
torchvision
timm
matplotlib
seaborn
psutil
numpy
scikit-learn
onnx
onnxruntime-gpu
onnxruntime-tools
pillow
